Split subscription bookkeeping out of LiveObjectMutableState
Preparation for implementing LiveObject and RealtimeObjects'
on(event:callback:) method.

This was largely done by Cursor.

Doing this in a bit of a rush and didn't have time to split out the
tests; will do in #81.
